+++ src/store/modules/loans.js
@@ -1,4 +1,5 @@
 import httpCommons from '../../http-commons'
+import { parseJwt } from '@/utils/utils.js';
 
 const apiName = 'loans'
 
@@ -17,12 +18,16 @@ const actions = {
     
     async getTotalLoanPerDepartment({commit}){
         var token = localStorage.getItem('token');
-        console.log('loan dept')
-        console.log(token)
-        var header = {
-            'Authorization' : 'Bearer '+token
-        }
-        const res = await httpCommons.get(apiName+'/totaLoanByDepartment', {headers : header})
+        const parseToken =parseJwt(token);
+        const user =parseToken.user
+        // console.log('loan dept')
+        // console.log(parseToken.user)
+        // var header = {
+        //     'Authorization' : 'Bearer '+token
+        // }
+        // const res = await httpCommons.get(apiName+'/totaLoanByDepartment', {headers : header})
+        const res = await httpCommons.get(apiName+'/totaLoanByDepartment/'+user.role )
+        console.log(res.data)
         commit('SET_DATA_LOAN_BY_DEPT', res.data)
     }
 }